fixed status code

+++ src/Pages/LoginPages/SignInPage/SignInPage.js
@@ -31,7 +31,7 @@ const SignIn = () => {
           password,
         }
       );
-      if (res.data.statusCode === '200') {
+      if (res && res.data && res.data.statusCode && res.data.statusCode === '200') {
         localStorage.clear();
         localStorage.setItem("user", JSON.stringify(res.data.data));
         localStorage.setItem("email", JSON.stringify(res.data.data?.email));
